added file via upload testing strategy, main files, and the real lic.
I had hardware problems so there are quit a few back ups. she is about an hour worth of button up stuff for her to go into trial. docuentation on this build and quite a few others can be found in the Life First App repo and the helix_ai_serer repo.